Add PositionsController for Lists and rename concern
