+++ app/controllers/lists_controller.rb
@@ -1,12 +1,9 @@
 class ListsController < AuthenticatedController
   include ActivityTracking
   include ProjectScoped
-  include ValidateMove
 
   before_action :set_current_board
-  before_action :set_list, only: [:edit, :update, :destroy, :move]
-  before_action :set_prev_item_and_next_item, only: :move
-  before_action :validate_move_params, only: :move
+  before_action :set_list, only: [:edit, :update, :destroy]
 
   def new
     @list = @board.lists.new
@@ -35,14 +32,6 @@ class ListsController < AuthenticatedController
     end
   end
 
-  def move
-    Board.move(@list, prev_item: @prev_item, next_item: @next_item)
-
-    track_updated(@list)
-
-    render json: @list
-  end
-
   def destroy
     if @list.destroy
       track_destroyed(@list)
@@ -56,23 +45,6 @@ class ListsController < AuthenticatedController
 
   def list_params
     params.require(:list).permit(:name)
-  end
-
-  def move_params
-    params.
-      permit(:id, :project_id, :board_id, :next_id, :prev_id)
-  end
-
-  def moveable_items
-    @board.lists
-  end
-
-  def moveable_item_name
-    'list'
-  end
-
-  def moveable_parent
-    @board
   end
 
   def set_current_board
