+++ app/views/cards/show.html.erb
@@ -22,7 +22,7 @@
       <div class="content-container">
         <div class="tab-content">
           <div class="tab-pane active" id="info-tab">
-            <% cache ['card-information-tab', @card, @card.list, @board] do %>
+            <% cache ['card-information-tab', @card, @board] do %>
               <div class="note-text-inner">
                 <h4 class="mb-4 header-underline">
                   <span class="text-truncate" title="<%= @card.name %>"><%= @card.name %></span>
